Added GPS Service, need fix to make it work in background Added gps permission requests to main activity After user clicks on submit btn, permission request should appear. After clicking on yes, activity should change and service should start.

+++ app/src/main/java/com/example/cmp354_covidtracer/DatabaseConnector.java
@@ -78,7 +78,7 @@ public class DatabaseConnector
 
    } // end method getAllContacts
 
-   public Cursor getExposureByUsersAndTimestamp(String userIdToSelect, String positiveIdToSelect, int timestampToSelect)
+   public Cursor getExposureByUsersAndTimestamp(String userIdToSelect, String positiveIdToSelect, long timestampToSelect)
    {
       //return database.query("contacts", new String[] {"_id", "name"},
       //	         null, null, null, null, "name"/*order by*/);
